fixed methods signatures
