+++ PerimeterXModule/DataContracts/Requests/RiskRequest.cs
@@ -1,9 +1,8 @@
-﻿using System;
-using System.Runtime.Serialization;
+﻿using System.Runtime.Serialization;
 
 namespace PerimeterX
 {
-	[DataContract]
+    [DataContract]
     [System.Obsolete("Use RiskRequestV2")]
 	public class RiskRequest
 	{
