+++ PerimeterXModule/DataContracts/Responses/RiskResponseV2.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.Runtime.Serialization;
+﻿using System.Runtime.Serialization;
 
 namespace PerimeterX
 {
