+++ PerimeterXModule/Internals/ActivityReporter.cs
