+++ PerimeterXModule/Internals/CookieDecoder.cs
