+++ PerimeterXModule/Internals/Helpers/PxConstants.cs
@@ -22,6 +22,7 @@ namespace PerimeterX
 
 		// Endpoints
 		public const string RISK_API_V2 = "/api/v2/risk";
+        public const string CAPTCHA_API_V1 = "/api/v1/risk/captcha";
 
         private static string GetAssemblyVersion()
         {
