+++ PerimeterXModule/Internals/Validators/Interfaces/IPXCaptchaValidator.cs
@@ -1,0 +1,7 @@
+﻿namespace PerimeterX
+{
+    interface IPXCaptchaValidator
+    {
+        bool CaptchaVerify(PxContext context);
+    }
+}
