+++ PerimeterXModule/Internals/Validators/Interfaces/IPXCookieValidator.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace PerimeterX
+{
+    interface IPXCookieValidator
+    {
+        RiskRequestReasonEnum CookieVerify(PxContext context, IPxCookie pxCookie);
+    }
+}
