+++ PerimeterXModule/Internals/Validators/Interfaces/IPXS2SValidator.cs
@@ -1,0 +1,7 @@
+﻿namespace PerimeterX
+{
+    public interface IPXS2SValidator
+    {
+        bool VerifyS2S(PxContext PxContext);
+    }
+}
