+++ PerimeterXModule/PerimeterXModule.csproj
@@ -73,7 +73,11 @@
     <Compile Include="DataContracts\Cookies\PxCookieV3.cs" />
     <Compile Include="DataContracts\Cookies\Helpers\DecodedCookieV1.cs" />
     <Compile Include="Internals\Helpers\PxConstants.cs" />
-    <Compile Include="Internals\Validators\IPXS2SValidator.cs" />
+    <Compile Include="Internals\Validators\Interfaces\IPXCookieValidator.cs" />
+    <Compile Include="Internals\Validators\Interfaces\IPXCaptchaValidator.cs" />
+    <Compile Include="Internals\Validators\Interfaces\IPXS2SValidator.cs" />
+    <Compile Include="Internals\Validators\PXCaptchaValidator.cs" />
+    <Compile Include="Internals\Validators\PXCookieValidator.cs" />
     <Compile Include="Properties\Resources.Designer.cs">
       <AutoGen>True</AutoGen>
       <DesignTime>True</DesignTime>
@@ -85,12 +89,12 @@
       <DependentUpon>Settings.settings</DependentUpon>
     </Compile>
     <Compile Include="Internals\PxContext.cs" />
-    <Compile Include="TemplateFactory.cs" />
+    <Compile Include="Internals\TemplateFactory.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="CookieDecoder.cs" />
+    <Compile Include="Internals\CookieDecoder.cs" />
     <Compile Include="PxModule.cs" />
     <Compile Include="PxModuleConfigurationSection.cs" />
-    <Compile Include="ActivityReporter.cs" />
+    <Compile Include="Internals\ActivityReporter.cs" />
     <Compile Include="DataContracts\Requests\RiskRequestHeader.cs" />
     <Compile Include="DataContracts\Requests\Request.cs" />
     <Compile Include="DataContracts\Requests\Additional.cs" />
@@ -107,8 +111,8 @@
     <Compile Include="DataContracts\Activities\ActivityDetails.cs" />
   </ItemGroup>
   <ItemGroup>
-    <EmbeddedResource Include="block.mustache" />
-    <EmbeddedResource Include="captcha.mustache" />
+    <EmbeddedResource Include="Internals\Templates\block.mustache" />
+    <EmbeddedResource Include="Internals\Templates\captcha.mustache" />
     <None Include="packages.config" />
     <None Include="Properties\Settings.settings">
       <Generator>SettingsSingleFileGenerator</Generator>
@@ -117,7 +121,7 @@
   </ItemGroup>
   <ItemGroup>
     <EmbeddedResource Include="Properties\Resources.resx">
-      <Generator>PublicResXFileCodeGenerator</Generator>
+      <Generator>ResXFileCodeGenerator</Generator>
       <LastGenOutput>Resources.Designer.cs</LastGenOutput>
     </EmbeddedResource>
   </ItemGroup>
