+++ PerimeterXModule/packages.config
@@ -1,0 +1,6 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Jil" version="2.15.0" targetFramework="net45" />
+  <package id="Nustache" version="1.16.0.4" targetFramework="net45" />
+  <package id="Sigil" version="4.7.0" targetFramework="net45" />
+</packages>
